refactor(types): move files inside types to a proper folder structure
