+++ @types/inline.response.d.ts
@@ -1,0 +1,5 @@
+export interface IInlineResponse<T> {
+  success: boolean;
+  data: T;
+  error: null | Record<string, unknown>;
+}
