+++ @types/model.objectid.d.ts
@@ -1,0 +1,3 @@
+import type { Document, Types } from 'mongoose';
+
+export type ID<T extends Document> = T['_id'] | Types.ObjectId | T;
