--- types/express.d.ts
@@ -1,7 +1,0 @@
-declare module 'http' {
-  interface IncomingHttpHeaders {
-    'x-secret-pass'?: string;
-    'x-local-dev-pass'?: string;
-    'x-lean-doc-request'?: string;
-  }
-}
