--- types/inline.response.d.ts
@@ -1,5 +1,0 @@
-export interface IInlineResponse<T> {
-  success: boolean;
-  data: T;
-  error: null | Record<string, unknown>;
-}
