--- types/model.objectid.d.ts
@@ -1,3 +1,0 @@
-import type { Document, Types } from 'mongoose';
-
-export type ID<T extends Document> = T['_id'] | Types.ObjectId | T;
